Remove deprecations (#264)
* remove goose.models

* remove Summary constructors

* remove GooseModels on liesel.model side

* remove obs and param helper functions

* remove GraphBuilder.transform calls in Var.transform tests

* User Var.transform instead of GraphBuilder.transform

* Use Var.transform instead of GraphBuilder.transform

* remove GraphBuilder.transform

* move transform tests to their own file

* remove goose.models

* remove Summary constructors

* remove GooseModels on liesel.model side

* remove obs and param helper functions

* remove GraphBuilder.transform calls in Var.transform tests

* User Var.transform instead of GraphBuilder.transform

* Use Var.transform instead of GraphBuilder.transform

* remove GraphBuilder.transform

* move transform tests to their own file

* removes deprecated function `stack_for_multi`

* turns warnings into errors and filters warnings

Adds filters for Specific DeprecationWarnings when they originate from dependencies

* Update 07-groups.qmd

* ignore matplotlib warning

* slight refactoring of Var.transform

* Update Var.transform tests

* remove dependencies deprecated and types-deprecated

* remove SamplingResult

* add link to tfp bijectors page to Var.transform docs

* update uv.lock

* Update nodes.py

* turn group deprecation into soft deprecation

* update test actions

* Revert "turn group deprecation into soft deprecation"

This reverts commit d11eb0fd9056a159fe4460bb211d7417774b397b.

* filter warnings

* update installation in actions

* filter warning in doctest

* ignore warning in pymc test

* Revert "filter warning in doctest" and "filter warnings"

This reverts commit edb883dc287f658001eee14393ecf2ac1b7e60a0.
This reverts commit d7e502757d1c6ba229afe868760306893a139734.

* remove deprecation from groups

---------
