+++ .github/workflows/pre-commit.yml
@@ -24,7 +24,7 @@ jobs:
           python-version-file: "pyproject.toml"
 
       - name: Install liesel
-        run: uv sync --locked --dev
+        run: uv sync --locked --group dev
 
       - name: Run pre-commit
         run: uv run pre-commit run -a
