+++ .github/workflows/pytest-pull.yml
@@ -25,7 +25,7 @@ jobs:
           cache-dependency-glob: "uv.lock"
 
       - name: Install liesel
-        run: uv sync --locked --dev
+        run: uv sync --locked --group dev --group pymc
 
       - name: Run pytest
         run: uv run pytest
