+++ tests/model/test_model.py
@@ -1,6 +1,5 @@
 import tempfile
 import typing
-import warnings
 from collections.abc import Generator
 from itertools import combinations
 from types import MappingProxyType
@@ -362,9 +361,8 @@ class TestModel:
         _, vars = model.copy_nodes_and_vars()
 
         gb.add(*vars.values())
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", (FutureWarning))
-            gb.transform(vars["x"])
+        vars["x"].transform()
+
         new_model = gb.build_model()
 
         assert "x_transformed" in new_model.vars
