+++ tests/model/test_var.py
@@ -1,11 +1,7 @@
-import pickle
 import typing
-import warnings
 
-import dill
 import jax
 import jax.numpy as jnp
-import numpy as np
 import pytest
 import tensorflow_probability.substrates.jax as tfp
 
@@ -566,301 +562,6 @@ class TestVarConstructors:
         loc = lsl.Var.new_value(1.0, name="loc")
         assert isinstance(loc, lsl.Var)
         assert loc.strong
-
-
-class TestVarTransform:
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_weak_var_with_distribtution_inst(self, name) -> None:
-        """
-        Tests transformation of a weak var with distribution when the bijector is passed
-        as an instance.
-        """
-        x = lsl.Var.new_value(jnp.linspace(0.1, 2, 5), name="all_x")
-        batch_index = lsl.Var.new_value(1, name="index")
-
-        x_batched = lsl.Var(
-            value=lsl.Calc(lambda i, x: x[i], batch_index, x),
-            distribution=lsl.Dist(tfp.distributions.Normal, loc=0.0, scale=1.0),
-            name="x_batched",
-        )
-
-        x_batched_transformed = x_batched.transform(tfp.bijectors.Exp(), name=name)
-
-        if name is None:
-            assert x_batched_transformed.name == f"{x_batched.name}_transformed"
-        else:
-            assert x_batched_transformed.name == name
-
-        assert x_batched_transformed.value == pytest.approx(jnp.log(x.value[1]))
-
-        batch_index.value = 2
-        x_batched_transformed.update()
-        x_batched.update()
-        assert x_batched_transformed.value == pytest.approx(jnp.log(x.value[2]))
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_weak_var_with_distribtution_class(self, name) -> None:
-        """
-        Tests transformation of a weak var with distribution when the bijector is passed
-        as a class.
-        """
-        x = lsl.Var.new_value(jnp.linspace(0.1, 2, 5), name="all_x")
-        batch_index = lsl.Var.new_value(1, name="index")
-
-        x_batched = lsl.Var(
-            value=lsl.Calc(lambda i, x: x[i], i=batch_index, x=x),
-            distribution=lsl.Dist(tfp.distributions.Normal, loc=0.0, scale=1.0),
-            name="x_batched",
-        )
-
-        x_batched_transformed = x_batched.transform(
-            tfp.bijectors.Scale, scale=2.0, name=name
-        )
-
-        if name is None:
-            assert x_batched_transformed.name == f"{x_batched.name}_transformed"
-        else:
-            assert x_batched_transformed.name == name
-
-        assert x_batched_transformed.value == pytest.approx(x.value[1] / 2.0)
-
-        batch_index.value = 2
-        x_batched_transformed.update()
-        x_batched.update()
-        assert x_batched_transformed.value == pytest.approx(x.value[2] / 2.0)
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_weak_var_with_bijector_instance(self, name) -> None:
-        tau = lsl.Var.new_param(10.0, name="tau")
-        tau_sqrt = lsl.Var.new_calc(jnp.sqrt, tau)
-        log_tau_sqrt = tau_sqrt.transform(tfp.bijectors.Exp(), name=name)
-
-        if name is None:
-            assert log_tau_sqrt.name == f"{tau_sqrt.name}_transformed"
-        else:
-            assert log_tau_sqrt.name == name
-
-        assert tau.value == pytest.approx(10.0)
-        assert tau_sqrt.value == pytest.approx(jnp.sqrt(10.0))
-        assert log_tau_sqrt.value == pytest.approx(jnp.log(jnp.sqrt(10.0)))
-
-        assert tau.strong
-        assert tau_sqrt.weak
-        assert log_tau_sqrt.weak
-        assert tau.parameter
-        assert not log_tau_sqrt.parameter
-        assert not tau_sqrt.parameter
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_weak_var_with_bijector_class(self, name) -> None:
-        tau = lsl.Var.new_param(10.0, name="tau")
-        tau_sqrt = lsl.Var.new_calc(jnp.sqrt, tau)
-
-        scale = lsl.Var.new_param(2.0, name="bijector_scale")
-        scaled_tau_sqrt = tau_sqrt.transform(
-            tfp.bijectors.Scale, scale=scale, name=name
-        )
-
-        if name is None:
-            assert scaled_tau_sqrt.name == f"{tau_sqrt.name}_transformed"
-        else:
-            assert scaled_tau_sqrt.name == name
-
-        assert tau.value == pytest.approx(10.0)
-        assert tau_sqrt.value == pytest.approx(jnp.sqrt(10.0))
-        assert scaled_tau_sqrt.value == pytest.approx(jnp.sqrt(10.0) / 2)
-
-        assert tau.strong
-        assert tau_sqrt.weak
-        assert scaled_tau_sqrt.weak
-        assert tau.parameter
-        assert not scaled_tau_sqrt.parameter
-        assert not tau_sqrt.parameter
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_without_dist_with_bijector_instance(self, name) -> None:
-        tau = lsl.Var.new_param(10.0, name="tau")
-        log_tau = tau.transform(tfp.bijectors.Exp(), name=name)
-
-        if name is None:
-            assert log_tau.name == f"{tau.name}_transformed"
-        else:
-            assert log_tau.name == name
-
-        assert tau.value == pytest.approx(10.0)
-        assert log_tau.value == pytest.approx(jnp.log(10.0))
-
-        assert tau.weak
-        assert log_tau.strong
-        assert not tau.parameter
-        assert log_tau.parameter
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_without_dist_with_bijector_class(self, name) -> None:
-        tau = lsl.Var.new_param(10.0, name="tau")
-
-        scale = lsl.Var.new_param(2.0, name="bijector_scale")
-        log_tau = tau.transform(tfp.bijectors.Scale, scale=scale, name=name)
-
-        if name is None:
-            assert log_tau.name == f"{tau.name}_transformed"
-        else:
-            assert log_tau.name == name
-
-        assert tau.value == pytest.approx(10.0)
-        assert log_tau.value == pytest.approx(5.0)
-
-        assert tau.weak
-        assert log_tau.strong
-        assert not tau.parameter
-        assert log_tau.parameter
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_instance(self, name) -> None:
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        log_tau = tau.transform(tfp.bijectors.Exp(), name=name)
-        tau.update()
-
-        if name is None:
-            assert log_tau.name == f"{tau.name}_transformed"
-        else:
-            assert log_tau.name == name
-
-        assert tau.weak
-        assert not log_tau.weak
-
-        assert tau.value == pytest.approx(np.exp(log_tau.value))
-        assert tau.value == pytest.approx(10.0)
-        assert log_tau.value == pytest.approx(np.log(10.0))
-
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", (FutureWarning))
-            log_tau_gb = lsl.GraphBuilder().transform(tau, tfp.bijectors.Exp())
-
-        assert tau.weak
-        assert not log_tau.weak
-
-        tau.update()
-        assert tau.value == pytest.approx(np.exp(log_tau.value))
-        assert tau.value == pytest.approx(10.0)
-        assert log_tau_gb.value == pytest.approx(np.log(10.0))
-
-        log_tau.dist_node.update()  # type: ignore
-        log_tau_gb.dist_node.update()  # type: ignore
-        assert log_tau.log_prob == pytest.approx(log_tau_gb.log_prob)
-
-    def test_transform_class_no_args(self) -> None:
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        with pytest.raises(ValueError):
-            tau.transform(tfp.bijectors.Exp)
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_class_with_args(self, name) -> None:
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        transformed_tau = tau.transform(
-            tfp.bijectors.Softplus, hinge_softness=lsl.Var(0.9), name=name
-        )
-        tau.update()
-
-        if name is None:
-            assert transformed_tau.name == f"{tau.name}_transformed"
-        else:
-            assert transformed_tau.name == name
-
-        bijector = tfp.bijectors.Softplus(hinge_softness=0.9)
-
-        assert tau.weak
-        assert not transformed_tau.weak
-
-        assert tau.value == pytest.approx(bijector.forward(transformed_tau.value))
-        assert tau.value == pytest.approx(10.0)
-        assert transformed_tau.value == pytest.approx(bijector.inverse(10.0))
-
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", (FutureWarning))
-            transformed_tau_gb = lsl.GraphBuilder().transform(
-                tau, tfp.bijectors.Softplus, hinge_softness=lsl.Var(0.9)
-            )
-
-        assert tau.weak
-        assert not transformed_tau.weak
-
-        tau.update()
-        assert tau.value == pytest.approx(bijector.forward(transformed_tau.value))
-        assert tau.value == pytest.approx(10.0)
-        assert transformed_tau_gb.value == pytest.approx(bijector.inverse(10.0))
-
-        transformed_tau.dist_node.update()  # type: ignore
-        transformed_tau_gb.dist_node.update()  # type: ignore
-        assert transformed_tau.log_prob == pytest.approx(transformed_tau_gb.log_prob)
-
-    @pytest.mark.parametrize("name", ("newname", None))
-    def test_transform_default(self, name) -> None:
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        log_tau = tau.transform(name=name)
-        tau.update()
-
-        if name is None:
-            assert log_tau.name == f"{tau.name}_transformed"
-        else:
-            assert log_tau.name == name
-
-        assert tau.weak
-        assert not log_tau.weak
-
-        assert tau.value == pytest.approx(np.exp(log_tau.value))
-        assert tau.value == pytest.approx(10.0)
-        assert log_tau.value == pytest.approx(np.log(10.0))
-
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", (FutureWarning))
-            log_tau_gb = lsl.GraphBuilder().transform(tau)
-
-        assert tau.weak
-        assert not log_tau.weak
-
-        tau.update()
-        assert tau.value == pytest.approx(np.exp(log_tau.value))
-        assert tau.value == pytest.approx(10.0)
-        assert log_tau_gb.value == pytest.approx(np.log(10.0))
-
-        log_tau.dist_node.update()  # type: ignore
-        log_tau_gb.dist_node.update()  # type: ignore
-        assert log_tau.log_prob == pytest.approx(log_tau_gb.log_prob)
-
-    def test_pickle_model_with_transformed_var(self, tmp_path):
-        prior = lsl.Dist(tfp.distributions.HalfCauchy, loc=0.0, scale=25.0)
-        tau = lsl.Var(10.0, prior, name="tau")
-        _ = tau.transform()
-
-        model = lsl.Model([tau])
-
-        with pytest.raises(AttributeError):
-            with open(tmp_path / "model.pkl", "wb") as f:
-                pickle.dump(model, f)
-
-            with open(tmp_path / "model.pkl", "rb") as f:
-                model2 = pickle.load(f)
-
-        with open(tmp_path / "model.pkl", "wb") as f:
-            dill.dump(model, f)
-
-        with open(tmp_path / "model.pkl", "rb") as f:
-            model2 = dill.load(f)
-
-        assert len(model2.vars) == len(model.vars)
-        assert len(model2.nodes) == len(model.nodes)
-        assert model2.log_prob == pytest.approx(model.log_prob)
 
 
 class TestVarPredictions:
